Rename blacklistServiceConfig to BlackListServiceConfig
